% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credsubs.R
\name{run.shiny.calc}
\alias{run.shiny.calc}
\title{Run a calculator}
\usage{
run.shiny.calc(app.dir = system.file("shiny", "alzheimers", package =
  "credsubs"))
}
\arguments{
\item{app.dir}{A character string pointing to the application directory.}
}
\description{
Runs the specified calculator using the \code{shiny} package.
Calculators must be built using \code{build.shiny.calc}.
}
\details{
The \code{app.dir} argument need not exactly match the value recommended
by \code{build.shiny.calc()}, as long as it points to the correct directory.
For example, \code{"./calc/"}, \code{"calc/"}, and \code{"calc"} are
all equivalent. If no value is supplied, an example is run.
}
